Make web-simple resilient against database connection errors (#3080)
Previously, shortly after the database tests began to run, web-simple
would say that its connection to MySQL had been lost with many
occurrences of this error message:

  Server web-simple: DBD::mysql::st execute failed: Lost connection to
  MySQL server during query at ...

Then for the remainder of the test it would print this error message:

  Server web-simple: DBD::mysql::st execute failed: MySQL server has
  gone away at ...

I think the database handle (the `$dbh` variable) would become corrupt
somehow and the application would never attempt to reconnect.  Now,
instead of using `DBI->connect` once we use `DBI->connect_cached` each
time we want a connection, which should reuse one connection as long as
it's active and reconnect when it breaks.  And we use
`$dbh->prepare_cached(...)` instead of `$dbh->prepare(...)` so that we
don't necessarily create a new statement on each request.

As a minor change, the database tests are now actually using the id
column values they read from the world table instead of ignoring the ids.
I'm not sure if this is a true requirement of the database tests, but it
feels like it should be.

The indentation in `app.pl` was inconsistent so many of the line changes
in this commit are indentation fixes.

I updated web-simple's cpan dependencies because why not.