Update hexagon dependencies (#3103)
Most importantly, update the kotlin and gradle versions to versions that
should be compatible with Java 9.

I added the javax.activation dependency to fix an issue running on Java
9, as the comment in build.gradle explains.  There is probably another
way of accomplishing this, like with "--add-modules java.activation" or
whatever it is you can do with the java command, but I don't know what
the equivalent of that incantation is in gradle/kotlin land.

I could not figure out how to upgrade the framework itself to the latest
version.  It seems that some of the framework's submodules have been
renamed.