Update pronghorn's dependencies
Update kotlin and gradle for Java 9 compatibility.

The change to not specify a Java version in httl.properties is also for
Java 9 compatibility.  It's not clear what the advantage of specifying
the Java version here would be.  The property seems to default to
whatever the currently-running version of Java is, which seems like what
we want.  See: https://github.com/httl/httl/blob/a7c71a953294954b539c25fefe17598cbbe1c571/httl/src/main/resources/httl-default.properties#L51

I updated the jsoniter and javassist dependencies as well because why not.