Update http4k's dependencies
Most importantly, update the kotlin and gradle versions for compatibility
with Java 9.

The original gradle-one-jar plugin is not maintained and is not
compatible with gradle 4, so we're now using a fork.

Other dependencies were updated just because.

It seems like there is a newer version of http4k itself available, 3.1.3.
I didn't attempt that upgrade because I assumed it would require
non-trivial changes to the test implementation.