Update ninja's dependencies, minor cleanup
This makes ninja-standalone run on both Java 8 and Java 9.  Before, it
seemed like the old version of Hibernate was causing problems on Java 9.
The part that initialized the HikariCP pool was throwing an exception,
something related to the pool max and min sizes being set out of order
and failing a bounds check.