Add dependencies to get rid of errors in Java 9 for utow-jersey-c3p0
Without the javax.activation and javax.xml.bind dependencies, I was
seeing stack traces in the logs, though it still seemed to pass the
tests.  The stack traces were similar to those I've seen in other
frameworks when run on Java 9, so I added the same dependencies for
fixing them just in case.

Also add the "useSSL=false" to the MySQL connection string to remove a
bunch of warning messages.