Update SBT from 0.13.9 to 1.0.4
This makes several Scala frameworks compatible with Java 9.

Add the build.properties specifying an old SBT version to blaze and
colossus because without that file, and with the SBT 1.0.4 upgrade, a
build.properties is auto-generated with `sbt.version=1.0.4`, and this
causes problems for the sbt-onejar plugin.  Having a build.properties
file puts blaze and colossus in line with all the other Scala frameworks,
which had that file already.