Remove the .gitattributes
I was wondering why, as of recently, I was having to manually convert
all my line endings each time I wanted to commit something to TFB.  It
looks like this (relatively new) .gitattributes file is to blame.  This
file came in with a commit to add a new framework back in November.  It
doesn't seem like it was addressing any real issues we were having.