Remove the permeagility framework
The database tests were not communicating with the database server, so
it was violating the test requirements.  We noticed this in issue #2346,
but we closed that issue prematurely without fixing the problem.

The framework seems to require OrientDB.  Personally, I don't think it's
worthwhile to install new, relatively obscure database engines for the
sake of relatively obscure frameworks.  If a web framework comes with a
core requirement like this that is at odds with our setup, I think it's
reasonable to say "this isn't a good fit" rather than trying to squish it
in.

The JSON test is failing, so without the (invalid) database tests, the
only one left is plaintext.  In our last preview, R15P3, this framework
had the lowest non-zero plaintext throughput at 283 requests per second,
with 2,621 errors, which I think works out to 175 errors per second, or
a 40% request error rate.

On Java 9, all the tests fail completely.

Taken together, the above points lead me to conclude that this code isn't
worth maintaining right now.

I'm not opposed to someone else re-adding permeagility later, but it
would need some changes.  If re-added, the new implementation should:

 - Pass the JSON and plaintext tests with reasonable error rates.
 - Either (a) not implement the DB tests, or (b) connect to one of the
   existing DBs (MySQL, PostgreSQL, MongoDB).
 - Be compatible with Java 9.