Remove the beyondj framework (#3135)
* Remove the beyondj framework

This one has been failing all tests for a while.  I don't think anyone is
going to fix it.  I've tried to fix it a few times, and I was never able
to.  The framework itself doesn't appear to be maintained anymore, so I
don't expect the framework author to come fix it.  And the framework is
very obscure and the implementation looks really complicated, so I don't
think any users are going to come fix it either.

I have no objections to beyondj being re-added in the future if someone
can write an implementation that works.

* Remove reference to now-deleted beyondj from .travis.yml