Set useSSL=false in more MySQL connection strings (#3148)
I found these connection strings by searching for "cachePrepStmts" and
ignoring the ones that were actually using PostgreSQL.  So if there are
MySQL connection strings that don't set cachePrepStmts, I missed them.

I didn't test the performance impact in all of these frameworks, but for
the three I did test, there was a notable positive benefit from setting
useSSL=false compared to leaving it unspecified.  When it's unspecified,
we sometimes see warning messages like this in the logs:

  Establishing SSL connection without server's identity verification is
  not recommended. According to MySQL 5.5.45+, 5.6.26+ and 5.7.6+
  requirements SSL connection must be established by default if explicit
  option isn't set. For compliance with existing applications not using
  SSL the verifyServerCertificate property is set to 'false'. You need
  either to explicitly disable SSL by setting useSSL=false, or set
  useSSL=true and provide truststore for server certificate verification.

That seems to say that if we leave useSSL unspecified, then SSL will be
used.  Maybe that explains the performance difference.