Disable in-memory caching of worlds and fortunes for minijax
We noticed minijax's multi-query results/sec were 4x higher than any
other framework's.  Testing locally, we saw that the multi-query and
fortunes pages both did not reflect updates made directly in the database
while the application server was running.  In other words, if we changed
the randomNumber of a world or the message of a fortune in the database
directly, the application would keep showing the old randomNumber or
message.

Adding the shared-cache-mode=NONE setting to persistence.xml appears to
fix this issue.