Use OpenJDK instead of the Oracle JDK
The oracle-java8-installer package is currently failing with a 404 error.
We found workarounds to fix the installer but they seem more complicated
than using OpenJDK instead.

We don't think that anything in TFB actually depends on something that's
in the Oracle JDK and not in OpenJDK such that test implementations would
start to fail.  If there are test implementations like that, we should
probably fix them so that they aren't like that.

We're asserting that there will be no difference in performance.

Looking ahead to the future, Oracle provides OpenJDK 9 in .tar.gz form
(at http://jdk.java.net/9/), which we would prefer to use instead of apt.
We're going to upgrade Java versions eventually, so we'd probably switch
to OpenJDK then for the .tar.gz convenience even if we didn't switch now.