Attempt to fix resetexpress build on Travis
The restexpress-mysql-raw test is failing with exceptions that look like
it's unable to connect to the mongodb database.  I didn't dive into the
implementation, but I'm guessing that it always connects to both
databases, but that depending on which setup file it's started with, it
only declares its dependency on one of the databases, and so the other
database isn't guaranteed to be available.  This change has it declare a
dependency on both databases always.