Add the facil.io framework
The [facil.io micro framework](http://facil.io) is somewhat oriented
for Websockets and pub/sub services, but it’s also used for RESTful
APIs and it powers the Ruby iodine server (an application server).

I think it would be great to add it to the benchmarking suite as well.