Added a lot of logging and fixed a bug
The logging will be removed once we figure out bugs.

The bug had to do with looping until there were fewer
than 2 child processes, and it turns out that we only
create a single child process from TFBReaper in the
situation where all child processes were killed in
the last iteration, and we call findChilds again.