Remove the officefloor framework
All of officefloor's tests have been failing in the ServerCentral
environment for a long time now.  We've tried to fix them more than once
but never succeeded.  On the previous attempt I filed this issue in the
framework's GitHub repository:

  https://github.com/sagenschneider/OfficeFloor/issues/8

Since I didn't get a response there and since there hasn't been any other
activity in that GitHub repository for the past few months, I'm assuming
that it's unlikely the framework's maintainer will come help us.

I would not be opposed to someone else re-adding the officefloor tests
later if we have some reason to believe it would start working on
ServerCentral (or whatever our dedicated hardware environment is at that
time).