Fix the database host name for vertx-web (#3204)
The setup script was trying to replace "localhost" with the IP of the
actual database server, but it wasn't doing anything because the config
file had "127.0.0.1", not "localhost".  But there's no need to do any
replacing; the "TFB-database" name exists for exactly this purpose.

In local testing, this change fixes all the database-related tests for
vertx-web.