Enable console logging for gemini, remove unused files (#3203)
Before, we had disabled gemini's logging completely, so if there was a
problem (such as: can't connect to the database and throws an exception
on startup) it wouldn't show up anywhere.  With this change, errors like
that should show up.

Starting resin with "resinctl console" instead of "resinctl start" means
the stdout from the gemini Java application will actually show up in
TFB's console output and get captured in the results directory.  Resin's
documentation suggests that "start" should be the mode used in
production, but that happens to be inconvenient for us in the context of
TFB since stdout is the only kind of logging we can depend on.  We're
asserting that there will be no difference in performance from "console"
as opposed to "start".

I removed all the unused .conf files for Gemini.  In the remaining used
.conf file, I enabled logging to stdout and changed the log threshold to
the recommended value.  I removed the "RequestLog*" settings because that
feature was removed from gemini.

I removed the Eclipse project files and the setup.sh script because those
aren't required by TFB.  They're leftovers from the original gemini
project generator and we don't intend for anyone to actually use them, if
they even still work.