Try to fix valum test timeout
In the most recent run on ServerCentral, valum had this error message in
the "completed" map of results.json:

  __run_test timeout (=7200 seconds)

When I ran it locally, it would pause strangely.  For example, when
verifying the JSON output, I would see this in the console:

  Accessing URL http://TFB-server:3003/json:
  Response (trimmed to 40 bytes): "{"message":"Hello, World!"}"

...but then it would stop there and nothing else would be printed.  If
I ran "curl localhost:3003" in another ssh tab, it would un-pause the
test and proceed the next one, where it would stop again.

I have no explanation for what was going on there, but I assume the same
thing was responsible for the timeouts on ServerCentral.

I noticed that the valum implementation code wrapped everything in an
"int main" function whereas none of the examples on the valum GitHub
repository did that.  Removing that wrapping function caused the weird
pauses to go away, making valum pass all the tests locally.