Fix silicon json test output, fix to fix new installations
The output from silicon's json endpoint was missing quotes around a
value, so it was failing the tests:

  {"message":Hello, World!}

I changed the code to look more like the hello world example on the
silicon website, and that made it pass tests locally.
http://siliconframework.org/docs/hello_world.html

Before I could do that, I had to tweak the silicon installer to depend
on a later version of gcc, or else the silicon installer would fail and
so would all of the test types.