Link the framework test to the updated facil.io version
Reviewing the testing logs exposed the fact that facil.io detected 80
CPU cores on the host machine (?!)… which invoked 80 forks.

This large number of forks was never attempted before and it exposed a
number of issues which were fixed in the beta.2 release.

This release also includes other minor updates and fixes which are
probably irrelevant (such as updated to the pub/sub and a fix for the
Redis engine etc’).