Downgrade hhvm from 3.24 to 3.21 to fix segfaults (#3233)
This should fix the hhvm variants of codeigniter, yii2, and laravel.

In commit 903997f3be7dbe3c2fba93accbcb0023063c1167 I worked around this
segfault for the no-framework hhvm tests by removing a "SET NAMES"
command we were issuing to MySQL and by specifying utf8 as the charset a
different way.  For these other frameworks though, the same strategy
isn't viable because the call to set the charset is happening deep in the
frameworks themselves, not in our test implementation code.