Upgrade to swift 4, repair vapor build
This fixes issue #3181.

I think we were running into this issue in the stable vapor release:
https://bugs.swift.org/browse/SR-5800

...which the comments claim is really about this issue:
https://bugs.swift.org/browse/SR-5802

...which was fixed in the main swift source repository:
https://github.com/apple/swift-package-manager/pull/1419

That fix happened pretty recently and doesn't seem to be included in the
latest stable releases (I don't see the PackageManager.swiftdoc file in
those releases), but it does seem to be fixed in the development snapshots
(I do see PackageManager.swiftdoc there).

With the development snapshot, the vapor build progresses much farther.
It eventually runs into an error compiling the vapor framework.  That is
an error in the outdated version of vapor we were using, and it was fixed
later:
https://github.com/vapor/vapor/commit/5a31fc8c7c6a54a8a98494795e0668666c341a33#diff-4a85e78227146848570c0c42d00d2f17

When I specify 2.4.4 as our version of vapor (the most recent release),
the build is successful and all the tests pass.

I renamed "swift3.sh" to "swift.sh" instead of "swift4.sh" because we
prefer to keep the language version out of these filenames unless we
can't avoid having multiple versions of the same language in the toolset.

It looked like vapor's setup-*.sh files were declaring unnecessary
database dependencies, so I cleaned those up.