Add index for world.id in mongodb
Some frameworks query for worlds by _id, and others by id.  It looks like
the ones that queried by id were being penalized heavily because we had
no index on that field.  In local testing, the RPS of nodejs-mongodb in
the db test went up by 8x (!) after I added the index.

Use createIndex instead of ensureIndex because, according to mongodb's
documentation, ensureIndex is:

  Deprecated since version 3.0.0: db.collection.ensureIndex() is now an
  alias for db.collection.createIndex().

Remove comment with a link to mongodb optimization because it doesn't
seem to say anything related to what we're doing in this file.