Tune JSON route and update facil.io
This update continues the lessons learned from the last round of tests
and updates both the facil.io version and the JSON path.

This experience was very enlightening and evoked and updated set of
default values for the facil.io framework (reflected in the updated
version).

This version includes a fixed CPU core limiter as a library default
(the last update did not properly limit the CPU cores) as well as a
connection capacity limiter and other memory allocation improvements.

As part of the efforts to minimize memory allocation values, the
`fiobj_hash_new2` function was introduced. This allows very small (or
very large) hash tables to hint at the requested capacity (minimizing
allocations when possible).