Fix web2py-optimized updates test implementation
It was failing verification before because it wasn't actually making all
the expected updates in the database.  I think what was happening was
it was building up a single long query containing multiple statements
like this:

  "UPDATE World SET randomNumber=4 WHERE id=5;UPDATE World SET randomNumber=1 WHERE id=2;..."

But the database driver was ignoring all but the first statement.  As far
as I know, most drivers prevent you from executing multiple statements in
one by default as a SQL injection countermeasure.

The revised implementation executes each UPDATE individually.  We've
learned from previous experiments in this project (with other
frameworks) that there isn't any way to execute a batch update against
the MySQL database that's faster than many individual updates.