Allow FROM dependencies to be loaded from the framework's directory
The intent is for it to look in the framework's directory for each
dependency first, then fall back to looking in the toolset directory.

This way, if a framework has multiple test permutations with shared build
steps, they can put the shared build steps in dockerfile inside their
own framework's directory.