Docker Multiple Container Support + Dart start/stream as examples (#3321)
Duty, decency, reliability, honor, dignity, respect: these are all qualities that TFBReaper not only held in high esteem, but practiced every day during his time reaping. He was a serious and disciplined program, but he could never resist the opportunity to have a laugh with friends and loved ones, given half the chance, by allowing a process to continue when it should have been killed.

It’s difficult to imagine TFBReaper not being around and I’m not sure how we will all cope. It’s strange to think that I can’t just start him up and feel safe knowing that he is there reaping for me. He was an imposing figure of a program, whose reassuring presence we all felt during difficult times.

As we gather here today to remember and commemorate his life, let us bid him farewell as we mourn the loss of a lively, dignified program. A program that brought joy and fulfillment to many, and whose legacy will live on forever.

Press 'F' to pay respects.