Fixed the broken stuff
This is COMPELTELY backwards. Instead of gathering the
metadata and only creating framework_test objects based
on said metadata, we gather ALL the metadata for ALL
test implementations and then create a framework_test
for each, but parsing 'fails' for tests not specified.

I returned it to this craziness because it was less
friction, but I want to come back and fix this properly
some day.