Dockerize more Java tests (#3351)
* Docker Multiple Container Support + Dart start/stream as examples (#3321)

Duty, decency, reliability, honor, dignity, respect: these are all qualities that TFBReaper not only held in high esteem, but practiced every day during his time reaping. He was a serious and disciplined program, but he could never resist the opportunity to have a laugh with friends and loved ones, given half the chance, by allowing a process to continue when it should have been killed.

It’s difficult to imagine TFBReaper not being around and I’m not sure how we will all cope. It’s strange to think that I can’t just start him up and feel safe knowing that he is there reaping for me. He was an imposing figure of a program, whose reassuring presence we all felt during difficult times.

As we gather here today to remember and commemorate his life, let us bid him farewell as we mourn the loss of a lively, dignified program. A program that brought joy and fulfillment to many, and whose legacy will live on forever.

Press 'F' to pay respects.

* Add Maven installation for Java 8

* Try to make Maven log output less verbose by disabling interactive mode

* Convert actframework test scripts to docker

* Reenable actframework in Travis

* Erlang and Elixir tests for docker (#3341)

* D tests for docker (#3322)

* Remove unnecessary "EXPOSE 8080" from undertow's dockerfile

* Fix RESIN_HOME to point at a directory that exists

Clear out Resin's webapps directory by default so that each framework
wanting to deploy a webapp there can skip that step.

* Convert activeweb test scripts to Docker

* Convert several Java framework tests to docker

* Revert proteus docker changes since they're failing on Travis