Broke everything (#3350)
* Broke everything

* Fixin' bugs and removing annoying comments

* Fixed the broken stuff

This is COMPELTELY backwards. Instead of gathering the
metadata and only creating framework_test objects based
on said metadata, we gather ALL the metadata for ALL
test implementations and then create a framework_test
for each, but parsing 'fails' for tests not specified.

I returned it to this craziness because it was less
friction, but I want to come back and fix this properly
some day.

* Linter

* Fixed formatting across python files
* Fixed namespacing
* Fixed executable PYTHONPATH

* Benchmarker.py now benchmarks

* Better filename convention

Or at least it's more consistent

* docker-helper now in charge of docker

* Capitals... why!?

* More linting

* More linting

* More linting

* Restructuring everything

With this I deleted the windows and sqlserver folders.
Docker should get us closer to running on windows, and
all the things in these folders were artifacts of the
past (read: gibberish). It's in git if we ever need it.

* Merge conflicts

* Fixed a bug

* Hosts are only needed in the docker containers

* print

* Fixed the update test verification

* Fixes a bug with ctrl+c