Port a few PHP tests to docker (#3451)
* Port a few PHP tests to docker

There are both php5 and php7 tests here, so hopefully that makes porting
the rest easier.  I guess the other big category (missing here) is HHVM.

For some reason I had to use "sleep infinity" as the last line of these
bash scripts instead of "wait", which I had used for the same purpose
for other frameworks.  I don't know why that is.

The zend test prints out a lot of errors but it still passes verification
locally.

* Simplify PHP dockerfiles