[ci skip] WIP: Dockerify everything
This is the proof of concept work that moves everything we do into
Docker. The toolset is invoked in a container, wrk is run from a
container, the framework implementations are run from within a
container, and the databases are run in a container.

Currently, this commit will only work locally with Docker, but
eventually the suite will support multi-machine setups (like our
production environments suitable for benchmarking).

Requirements to run this locally:
1) Docker installed
2) Build the toolset image:
  docker build -t tfb - < Dockerfile
3) Build the wrk image:
  docker build -t wrk - < toolset/setup/docker/wrk.dockerfile
4) Run the toolset:
  docker run --network=host -v [your docker socket]:/var/run/docker.sock --mount type=bind,source=[abspath to your FrameworkBenchmarksDir],target=/FrameworkBenchmarks tfb --mode verify --test gemini