Fixed ctrl+c and SIGTERM handling
Also, updated README.md to capture notes about how to run the toolset as
well as explanations to what the arcane flags do and why.

Lastly, removed tfb.sh because it would cause the suite to become
unresponsive to ctrl+c and SIGTERM. Toolset MUST be run from Docker
directly.