[ci skip] toolset: don't exit with an error status if the run "completes" (#3567)
The previous full run on Citrine exited with an error status code even
though it completed and was what we'd call a "successful" run.  It
probably ran into issues testing a couple of frameworks along the way,
which we assume that it logged, but the toolset didn't crash, and it
wasn't invoking with invalid arguments or anything.  So the command to
run the toolset should be considered a success.