[ci skip] remove redkale
We had to block the redkale contributor on GitHub for not listening to
our feedback on PRs.  Blocking them means they won't be able to update
the redkale implementation any more.  We think it would be unfair to
keep what they've added up to this point while not letting them update
anything, given that they are the creators of the framework.