Keep undertow and undertow-postgresql, remove the other variants
In general I think it's better for our test implementations to show one
good way of doing things with a given framework than it is to lay out
several options.  Based on previous results and the way the code turns
out, using postgresql with the official JDBC client is (in my opinion)
better than the other options for database access with this framework.