Make HTTP header verification more strict (#3643)
* Make HTTP header verification more strict

It has long been a *requirement* for responses to include these headers:
 - Server
 - Date
 - Content-Type
 - either Content-Length or Transfer-Encoding

Now we are enforcing that requirement instead of emitting a warning.

Frameworks missing these headers will no longer be benchmarked (until we
fix them).  I bet there is at least one framework that will fail because
of this.  I'm sure I've seen warnings about a missing Server header
before.

Also, do not emit a warning for frameworks that include ";charset=utf-8"
on plaintext and json responses.  Many frameworks do this by default and
that behavior isn't a problem that we want anyone to fix.

* [ci skip] Fix toolset exception when Content-Type is missing

* Try fixing several header verification failures

* Add validation for Date header format

* Fix more header verification failures

* Remove test code from ninja-stanalone, update dependencies

Having the test dependencies causes ninja-standalone to take much longer
to build.  And we don't encourage putting test code in here in general.
The implementations ARE the test code.

* Make invalid Date header format a warning, not a failure

I can't fix octane's or ninja-standalone's date formats.

* Use octane source files in implementation, remove unused files

Previously it was ignoring all of our source files and only running code
found in the octane GitHub repository.

* Try harder to make aspnet include charset=utf-8 in fortunes

* Fix wrong content type for yii2 plaintext, was html

* Fix Date headers for crystal frameworks

* Fix up aspnet content type problems finally, hopefully