[ci skip] Log Less
Gemini got hit pretty bad in the last few plaintext benchmarks on
Citrine, and we suspect that it was due to removing the resin.xml which
configured logging.

This commit attempts to turn off access logging entirely, although in
prior builds, the `access-log` directive was not required; may remove in
a future commit.