Don't consider a server up after 5xx (#3709)
Without this change, we consider a server ready for verification if it
responds to a curl command with any HTTP response, even if that response
indicates a failure like 502 Bad Gateway.  (This particular response can
happen with frameworks that use nginx, when nginx has started but the
back end server isn't ready yet.)  With this change, we wait for a
successful response.