[ci skip] try to free up disk space between continuous runs (#3714)
We rely on docker's built-in caching to save on build times.  We never
clear out this cache, so assuming people keep making changes to the
test implementations, we'll eventually run out of disk space.  This
happened most recently in the Citrine environment after 6 or so runs.

With this change, we'll try to clear out docker's cache when we notice
that the available disk space is uncomfortably low.  Currently the script
assumes that 20G is enough space to handle a full run.