[ci skip] Remove source_code files, count all files instead
Most of the source_code files we have are wrong.  They are usually not
kept up to date when the test implementation is refactored.  There's no
motivation to do so.

This change removes the source_code files and instead tells `cloc` to
count all files in the test directory.

This should be a more reasonable thing to do now that we're using docker.
Before docker, we'd build or download code in the test directories right
next to the source, which would have thrown off the counts by a lot.
Now these directories should basically remain in their clean state.

Separately, I think we should re-examine whether counting lines of code
is a reasonable thing to be doing at all, and if boiling it down to just
one number is reasonable, and if `cloc` is the right tool to use.  For
now we're still using `cloc` and the output format in results.json is
unchanged.