[ci skip] Add SQL Server (on linux) as a database type
I added [ci skip] because this requires a new techempower/tfb image, so
the travis build would fail I think.

I don't know whether this will work out because of licensing.  See the
places where I put "ACCEPT_EULA=Y" in the dockerfiles.  I don't know what
EULA that refers to and it's possible we won't be able to accept it.
I put this together anyway just to see if it's technically possible (it
is) and to get review feedback.

I added a variant of gemini that uses SQL Server and passes all tests.

I did not measure performance at all or do any performance tuning on the
database side.

Fixes #3786