Update the configurations of the play framework to their current latest version: 2.6.15
Also update the dockerfiles to use a more recent base, allowing use of Scala 2.12.6 and sbt 1.1.5

Apart from that, I noticed some docker files providing extra JVM options. Considering these are all using the same framework, I feel either all dockerfiles should use it or non of them. Otherwise, the comparison between them is off.

Lastly, I added some code to the dockerfile to properly disable the http server that's not being used in the build that will be done by the dockerfile.
For instance: when running netty server, it's perfectly fine to disable the AkkaHttpPlugin. And vice versa. This prevents their jars from being put on the classpath.