H2O: Optimize the cached queries test (#3904)
The main idea is to use several caches behind an interface that makes
them appear as a single one. Every cache operation chooses one of
them based on the key hash value, and while all subcaches are
multithreaded, the interface in front of them does not modify
anything, so it does not contain any critical sections. As a result,
if several threads access the cache, contention will be minimized,
provided that most of the time they use keys that map to different
subcaches.