Re-apply the undertow plaintext ByteBuffer optimization
Undertow's relative placement in the plaintext test dropped after
changing this handler to use a regular string.  This coincided with our
move to docker and a switch to new benchmarking hardware, so there are
certainly other variables in the mix, but it's reasonable to assume the
ByteBuffer->string change is partially to blame.  Also IIRC it was the
Undertow authors themselves who originally wrote the ByteBuffer
optimization, so it probably doesn't make sense for me to second guess
them.